[Fix] Fix mmpose deploy bug and update CI/CD (#70)
1. update Dockerfile to cuda-11.4 and mmcv-full-1.6.1
2. build a docker image and test it on a GPU machine
3. fix mmpose deploy bug when there's more than one bbox in an image
4. remove `np.float` which is no longer supported after numpy-1.24.0

+++ xrmocap/ops/top_down_association/identity_tracking/keypoints_distance_tracking.py
@@ -9,7 +9,7 @@ from .base_tracking import BaseTracking
 class KeypointsDistanceTracking(BaseTracking):
 
     def __init__(self,
-                 tracking_distance: np.float,
+                 tracking_distance: np.float32,
                  tracking_kps3d_convention: str,
                  tracking_kps3d_name: List[str],
                  verbose: bool = False,
@@ -18,7 +18,7 @@ class KeypointsDistanceTracking(BaseTracking):
         method assigns identities of the closest person in last frame.
 
         Args:
-            tracking_distance (np.float):
+            tracking_distance (np.float32):
                 The distance considered to be the same kps3d.
             tracking_kps3d_convention (str): The convention of kps3d.
             tracking_kps3d_name (List[str]): The name of the tracked kps3d.
